Move ReqWebMessage to be next to its servlet
