+++ com.ibm.streamsx.inet/impl/java/src/com/ibm/streamsx/inet/rest/ops/ReqHandlerInterface.java
@@ -1,4 +1,7 @@
 package com.ibm.streamsx.inet.rest.ops;
+
+import com.ibm.streamsx.inet.rest.servlets.ReqWebMessage;
+
 /**
 * Licensed Materials - Property of IBM
 * Copyright IBM Corp. 2014
